+++ src/struct/pdm_morton.h
@@ -452,7 +452,7 @@ PDM_morton_ordered_build_rank_index
  int                      gmax_level,
  PDM_l_num_t              n_codes,
  const PDM_morton_code_t  ordered_code[],
- const int                weight[],
+ const double             weight[],
  PDM_morton_code_t        rank_index[],
  PDM_MPI_Comm             comm
  );
@@ -479,7 +479,7 @@ PDM_morton_build_rank_index(int                      dim,
                             int                      gmax_level,
                             PDM_l_num_t              n_codes,
                             const PDM_morton_code_t  code[],
-                            const int                weight[],
+                            const double             weight[],
                             const int                order[],
                             PDM_morton_code_t        rank_index[],
                             PDM_MPI_Comm             comm);
